Added Float, Boolean, and String getters in TriggerProcessor.Config plus bugfixes

+++ app/src/main/java/io/crayfis/android/camera/CFSensor.java
@@ -95,7 +95,7 @@ class CFSensor implements SensorEventListener {
     }
 
     boolean isFlat() {
-        Integer orientCutoff = CFConfig.getInstance().getQualTrigger().getInt("orientation");
+        Float orientCutoff = CFConfig.getInstance().getQualTrigger().getFloat("orientation");
         return orientCutoff == null || Math.abs(rotationMatrix[8]) >= orientCutoff;
     }
